+++ lib/hooks/useBlockReload/useBlockReload.tsx
@@ -1,0 +1,15 @@
+import { useEffect } from "react";
+
+export function useBlockReload(predicate: boolean) {
+  useEffect(() => {
+    if (predicate) {
+      window.onbeforeunload = () => true;
+    } else {
+      window.onbeforeunload = null;
+    }
+
+    return () => {
+      window.onbeforeunload = null;
+    };
+  }, [predicate]);
+}